shopping cart controller completed.

+++ src/main/java/com/group3/onlineShooping/controller/PaymentController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import javax.validation.Valid;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -75,7 +76,7 @@ public class PaymentController {
         product.setAvailable(true);
         product.setAvailableInStor(10);
         product.setDescription("this is description");
-        product.setPrice(232233);
+        product.setPrice((BigDecimal.valueOf(232233)));
         product.setProductNumber("12312312");
         product.setSummary("this is summary");
         product.setTitle("title");
